chore(app): :zap: add sw services and interceptor for loading, change paths and imports
