+++ src/app/components/base/header/home-button/home-button.module.ts
@@ -1,7 +1,7 @@
-import { WbHbModule } from './../../../../wb-hb.module';
 import { NgModule } from '@angular/core';
 import { HomeButtonComponent } from './home-button.component';
 import { RouterModule } from '@angular/router';
+import { WbHbModule } from 'src/app/wb-hb.module';
 
 @NgModule({
   declarations: [HomeButtonComponent],
