+++ src/app/components/base/loader/loader.component.html
@@ -1,4 +1,4 @@
-<div *ngIf="isLoading$ | async" class="overlay">
+<div *ngIf="isLoading | async" class="overlay">
   <div class="loader">
     <div></div>
     <div></div>
