--- src/app/components/base/loader/loader.service.ts
@@ -1,19 +1,0 @@
-import { Injectable } from '@angular/core';
-import { Subject } from 'rxjs';
-
-@Injectable({
-  providedIn: 'root',
-})
-export class LoaderService {
-  isLoadingSubject = new Subject<boolean>();
-
-  constructor() {}
-
-  show() {
-    this.isLoadingSubject.next(true);
-  }
-
-  hide() {
-    this.isLoadingSubject.next(false);
-  }
-}
