+++ src/app/components/game/character/sheet/actions-features-notes/actions-features-notes.module.ts
@@ -2,12 +2,12 @@ import { NotesModule } from './notes/notes.module';
 import { FeaturesModule } from './features/features.module';
 import { ActionsModule } from './actions/actions.module';
 import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
 import { ActionsFeaturesNotesComponent } from './actions-features-notes.component';
+import { WbHbModule } from 'src/app/wb-hb.module';
 
 @NgModule({
   declarations: [ActionsFeaturesNotesComponent],
   exports: [ActionsFeaturesNotesComponent],
-  imports: [CommonModule, ActionsModule, FeaturesModule, NotesModule],
+  imports: [WbHbModule, ActionsModule, FeaturesModule, NotesModule],
 })
 export class ActionsFeaturesNotesModule {}
