+++ src/app/components/game/character/sheet/armor-class-speed/armor-class-speed.module.ts
@@ -1,12 +1,12 @@
 import { SpeedModule } from './speed/speed.module';
 import { ArmorClassModule } from './armor-class/armor-class.module';
 import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
 import { ArmorClassSpeedComponent } from './armor-class-speed.component';
+import { WbHbModule } from 'src/app/wb-hb.module';
 
 @NgModule({
   declarations: [ArmorClassSpeedComponent],
   exports: [ArmorClassSpeedComponent],
-  imports: [CommonModule, ArmorClassModule, SpeedModule],
+  imports: [WbHbModule, ArmorClassModule, SpeedModule],
 })
 export class ArmorClassSpeedModule {}
