+++ src/app/components/game/character/sheet/armor-class-speed/armor-class/armor-class.module.ts
@@ -1,10 +1,10 @@
 import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
+import { WbHbModule } from 'src/app/wb-hb.module';
 import { ArmorClassComponent } from './armor-class.component';
 
 @NgModule({
   declarations: [ArmorClassComponent],
   exports: [ArmorClassComponent],
-  imports: [CommonModule],
+  imports: [WbHbModule],
 })
 export class ArmorClassModule {}
