+++ src/app/components/game/character/sheet/list/item/item.module.ts
@@ -1,10 +1,10 @@
 import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
+import { WbHbModule } from 'src/app/wb-hb.module';
 import { ItemComponent } from './item.component';
 
 @NgModule({
   declarations: [ItemComponent],
   exports: [ItemComponent],
-  imports: [CommonModule],
+  imports: [WbHbModule],
 })
 export class ItemModule {}
