+++ src/app/components/game/character/sheet/skills/skills.module.ts
@@ -1,11 +1,11 @@
 import { SkillModule } from './skill/skill.module';
 import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
 import { SkillsComponent } from './skills.component';
+import { WbHbModule } from 'src/app/wb-hb.module';
 
 @NgModule({
   declarations: [SkillsComponent],
   exports: [SkillsComponent],
-  imports: [CommonModule, SkillModule],
+  imports: [WbHbModule, SkillModule],
 })
 export class SkillsModule {}
