+++ src/app/components/game/details/details.module.ts
@@ -1,17 +1,16 @@
-import { SheetModule } from './../character/sheet/sheet.module';
 import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
 import { DetailsComponent } from './details.component';
 import { FormsModule } from '@angular/forms';
-import { ButtonModule } from '../../base/button/button.module';
-import { InputModule } from '../../base/input/input.module';
 import { CharacterModule } from '../character/character.module';
 import { DetailsRoutingModule } from './details-routing.module';
+import { WbHbModule } from 'src/app/wb-hb.module';
+import { ButtonModule } from '@base/button/button.module';
+import { InputModule } from '@base/input/input.module';
 
 @NgModule({
   declarations: [DetailsComponent],
   imports: [
-    CommonModule,
+    WbHbModule,
     DetailsRoutingModule,
     CharacterModule,
     InputModule,
