+++ src/app/components/home/home.component.ts
@@ -5,6 +5,4 @@ import { Component } from '@angular/core';
   templateUrl: './home.component.html',
   styleUrls: ['./home.component.scss'],
 })
-export class HomeComponent {
-  loading!: boolean;
-}
+export class HomeComponent {}
