+++ src/app/components/home/home.module.ts
@@ -1,6 +1,5 @@
 import { LoginModule } from './../login/login.module';
 import { NgModule } from '@angular/core';
-import { RouterModule } from '@angular/router';
 import { FormsModule } from '@angular/forms';
 
 import { WbHbModule } from 'src/app/wb-hb.module';
