+++ src/app/definitions/keys/index.ts
@@ -1,0 +1,3 @@
+export * from './keys.d';
+export * from './keys.character.d';
+export * from './keys.response.d';
