+++ src/app/definitions/keys/keys.d.ts
@@ -1,0 +1,21 @@
+/* Keys for utility purpouses */
+export type QueryParamsKeys = 'gameid' | 'type' | 'id' | 'pretty';
+export type DataObjectKeys = 'object' | 'desc';
+export type NotesKeys = 'text' | 'title';
+export type InputKeys =
+  | 'color'
+  | 'custom'
+  | 'date'
+  | 'datetime-local'
+  | 'email'
+  | 'month'
+  | 'number'
+  | 'password'
+  | 'search'
+  | 'tel'
+  | 'text'
+  | 'time'
+  | 'url'
+  | 'week';
+export type TypeValueKeys = 'type' | 'value' | 'icon' | 'align' | 'color';
+export type InputAdditionalFieldsKeys = 'prefix' | 'suffix' | 'hint';
