+++ src/app/definitions/types/index.ts
@@ -1,0 +1,3 @@
+export * from './types.d';
+export * from './types.character.d';
+export * from './types.response.d';
