+++ src/app/services/tokens/api.ts
@@ -1,0 +1,3 @@
+import { InjectionToken } from '@angular/core';
+
+export const API_URL = new InjectionToken('API_URL');
