+++ src/styles/styles.scss
@@ -1,5 +1,5 @@
 /* Importing Bootstrap SCSS file. */
-@import '~bootstrap/scss/bootstrap';
+@import 'bootstrap/scss/bootstrap';
 
 @import '_functions';
 @import '_scrollbar';
